fix(Calculator): 소수점 계산이 가능하도록 변경
- 소수점 계산이 가능하도록 로직을 변경합니다.
- 메서드를 분할 합니다.

+++ src/main/java/calculator/controller/CalculatorController.java
@@ -20,7 +20,7 @@ public class CalculatorController {
         calculatorOutputer.printIntro();
         String inputString = calculatorInputer.getInput();
 
-        int result = calculatorService.calculateSumString(inputString);
+        float result = calculatorService.calculateSumString(inputString);
 
         calculatorOutputer.printResult(result);
     }
