+++ src/main/java/calculator/view/CalculatorOutputer.java
@@ -6,7 +6,7 @@ public class CalculatorOutputer {
         System.out.println("덧셈할 문자열을 입력해주세요.");
     }
 
-    public void printResult(int result) {
+    public void printResult(float result) {
         System.out.println("결과 : " + result);
     }
 }
